Add mailing service, Add CRUD endpoints for user/wather-triggers

+++ src/user/user-weather-trigger/user-weather-trigger.controller.ts
@@ -1,8 +1,19 @@
-import { Body, Controller, Delete, Param, Post, Req } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  NotFoundException,
+  Param,
+  Patch,
+  Post,
+  Req,
+} from '@nestjs/common';
 import { Types } from 'mongoose';
 import { Roles } from 'src/auth/decorators/roles.decorator';
 import { Role } from 'src/auth/models/role.enum';
 import { CreateWeatherTriggerDto } from 'src/weather-trigger/dto/create-weather-trigger.dto';
+import { UpdateWeatherTriggerDto } from 'src/weather-trigger/dto/update-weather-trigger.dto';
 import {
   WeatherTrigger,
   WeatherTriggerDocument,
@@ -17,18 +28,27 @@ export class UserWeatherTriggerController {
     private weatherTriggerService: WeatherTriggerService,
     private userService: UserService,
   ) {}
+
+  @Get()
+  async findAllWeatherTriggers(@Req() request): Promise<WeatherTrigger[]> {
+    const userId = await this.userService.getIdByUserName(
+      request.payload.username,
+    );
+    return this.weatherTriggerService.findBySubscriberId(userId);
+  }
+
   @Post()
   async createWeatherTrigger(
     @Body() createWeatherTriggerDto: CreateWeatherTriggerDto,
     @Req() request,
   ): Promise<WeatherTrigger> {
+    const userId = await this.userService.getIdByUserName(
+      request.payload.username,
+    );
+    createWeatherTriggerDto.subscriberId = new Types.ObjectId(userId);
     const newTrigger = (await this.weatherTriggerService.create(
       createWeatherTriggerDto,
     )) as WeatherTriggerDocument;
-    await this.userService.addSubscriptionToUser(
-      newTrigger._id,
-      request.payload.username,
-    );
     return newTrigger;
   }
 
@@ -37,10 +57,28 @@ export class UserWeatherTriggerController {
     @Param('id') id: string,
     @Req() request,
   ): Promise<WeatherTrigger> {
-    this.userService.removeSubscriptionFromUser(
-      new Types.ObjectId(id),
+    const userId = await this.userService.getIdByUserName(
       request.payload.username,
     );
+    const userTriggers = await this.weatherTriggerService.findOne(id, userId);
+    if (userTriggers === null) {
+      throw new NotFoundException();
+    }
     return this.weatherTriggerService.remove(id);
+  }
+  @Patch(':id')
+  async patchWeatherTrigger(
+    @Param('id') id: string,
+    @Body() updateWeatherTriggerDto: UpdateWeatherTriggerDto,
+    @Req() request,
+  ) {
+    const userId = await this.userService.getIdByUserName(
+      request.payload.username,
+    );
+    const userTriggers = await this.weatherTriggerService.findOne(id, userId);
+    if (userTriggers === null) {
+      throw new NotFoundException();
+    }
+    return this.weatherTriggerService.update(id, updateWeatherTriggerDto);
   }
 }
